+++ internal/domain/usecase/organization_finder.go
@@ -13,6 +13,6 @@ func NewOrganizationFinder(organizations repository.Organizations) *Organization
 	return &OrganizationFinder{Organizations: organizations}
 }
 
-func (it *OrganizationFinder) FindById(id int) (*entity.Organization, error) {
+func (it *OrganizationFinder) FindById(id int64) (*entity.Organization, error) {
 	return it.Organizations.Find(id)
 }
